# docs/swagger.yaml
basePath: /
definitions:
...
  delivery_http.CommonError:
    properties:
      error:
        type: string
    type: object
...
  domain.AdditionalAgreement:
    properties:
      agreement_date:
        type: string
      agreement_end_date:
        type: string
      agreement_number:
        type: string
      amount:
        type: number
      approval_status:
        type: string
      archived_at:
        type: string
      compliance_comment:
        type: string
      compliance_decision:
        type: string
      compliance_reviewed_at:
        type: string
      compliance_reviewed_by:
        type: string
      contract_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      currency_control_comment:
        type: string
      currency_control_decision:
        type: string
      currency_control_reviewed_at:
        type: string
      currency_control_reviewed_by:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      delivery_date:
        type: string
      doc_type:
        type: string
      document_path:
        type: string
      extend_date_to:
        type: string
      foreign_amount:
        type: number
      foreign_currency:
        type: string
      id:
        type: integer
      rejection_reason:
        type: string
      remaining_amount:
        type: number
      return_date:
        type: string
      status:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8088
info:
  contact: {}
  description: API сервера для системы валютного контроля.
  title: Currency Control API
  version: "1.0"
paths:
...
  /api/branches/{id}/dashboard/companies/{company_id}/contracts/{contract_id}/additional-agreements:
    get:
      description: Возвращает список дополнительных соглашений по контракту.
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AdditionalAgreement'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Список доп. соглашений контракта
      tags:
      - AdditionalAgreements
    post:
      consumes:
      - multipart/form-data
      description: Создает доп. соглашение к контракту с поддержкой прикрепления файла
        PDF/Word.
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Номер доп. соглашения
        in: formData
        name: agreement_number
        type: string
      - description: Дата доп. соглашения (YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: agreement_date
        type: string
      - description: Предмет соглашения
        in: formData
        name: subject
        type: string
      - description: Срок поставки товара / оказания услуг (дата YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: delivery_date
        type: string
      - description: Срок возврата денежных средств (дата YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: return_date
        type: string
      - description: Сумма доп. соглашения
        in: formData
        name: amount
        type: number
      - description: Валюта доп. соглашения (например USD, EUR, TJS)
        in: formData
        name: currency
        type: string
      - description: Дата окончания доп. соглашения (YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: agreement_end_date
        type: string
      - description: Тип документа (additional_agreement, specification, appendix)
        in: formData
        name: doc_type
        type: string
      - description: Файл доп. соглашения (.pdf, .doc, .docx)
        in: formData
        name: document
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AdditionalAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Создать доп. соглашение (Карточка Доп. соглашения)
      tags:
      - AdditionalAgreements
  /api/branches/{id}/dashboard/companies/{company_id}/contracts/{contract_id}/additional-agreements/{agreement_id}:
    delete:
      consumes:
      - application/json
      description: 'Помещает дополнительное соглашение в корзину (soft delete). Доступно:
        Сотрудники валютного контроля, Комплаенс, Администратор.'
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      - description: ID доп. соглашения
        in: path
        name: agreement_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Удаление доп. соглашения (в корзину)
      tags:
      - AdditionalAgreements
    get:
      description: Возвращает полную информацию по карточке дополнительного соглашения
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      - description: ID доп. соглашения
        in: path
        name: agreement_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AdditionalAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Карточка доп. соглашения (получить по ID)
      tags:
      - AdditionalAgreements
    put:
      consumes:
      - multipart/form-data
      description: 'Редактирование дополнительного соглашения. Доступно: Операционный
        сотрудник (при отправке на доработку), Сотрудник валютного контроля (при наличии
        разрешения), Комплаенс, Администратор.'
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      - description: ID доп. соглашения
        in: path
        name: agreement_id
        required: true
        type: integer
      - description: Номер доп. соглашения
        in: formData
        name: agreement_number
        type: string
      - description: Дата доп. соглашения (YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: agreement_date
        type: string
      - description: Предмет соглашения
        in: formData
        name: subject
        type: string
      - description: Срок поставки товара (дата YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: delivery_date
        type: string
      - description: Срок возврата денежных средств (дата YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: return_date
        type: string
      - description: Сумма доп. соглашения
        in: formData
        name: amount
        type: number
      - description: Валюта доп. соглашения
        in: formData
        name: currency
        type: string
      - description: Дата окончания доп. соглашения (YYYY-MM-DD или DD.MM.YYYY)
        in: formData
        name: agreement_end_date
        type: string
      - description: Тип документа
        in: formData
        name: doc_type
        type: string
      - description: Новый PDF/Word документ (опционально)
        in: formData
        name: document
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AdditionalAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Редактирование доп. соглашения
      tags:
      - AdditionalAgreements
...
  /api/branches/{id}/dashboard/companies/{company_id}/contracts/{contract_id}/additional-agreements/{agreement_id}/restore:
    put:
      description: 'Переводит доп. соглашение из статуса archived обратно в active.
        Доступно: Сотрудники валютного контроля, Комплаенс, Администратор.'
      parameters:
      - description: ID филиала
        in: path
        name: id
        required: true
        type: integer
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: integer
      - description: ID контракта
        in: path
        name: contract_id
        required: true
        type: integer
      - description: ID доп. соглашения
        in: path
        name: agreement_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Разархивировать доп. соглашение (восстановить в active)
      tags:
      - Archive
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите Bearer access_token (полученный при /auth/login или /auth/refresh)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"